Split off split_frame.
