+++ src/zebras/__init__.py
@@ -28,5 +28,6 @@ __all__ = [
 
 from ._freq import Freq, freq
 from ._period import period, Period, periods, Periods
-from ._timeframe import time_frame, TimeFrame, split_frame, SplitFrame
+from ._split_Frame import split_frame, SplitFrame
+from ._timeframe import time_frame, TimeFrame
 from ._time_series import time_series, TimeSeries
